+++ classes/Configuration Types/Number.js
@@ -2,8 +2,7 @@
 
 const fs = require("fs-extra-promise");
 const path = require("path");
-const dataDir = require("../Config.js").dataDir;
-const guildConfs = require("../Config.js").guildConfs;
+
 
 /** The starting point for creating a Number configuration key. */
 class NumberConfig {
@@ -19,6 +18,8 @@ class NumberConfig {
     if (data.max) this.max = data.max;
     this.type = "Number";
     Object.defineProperty(this, "_id", { value: conf._id });
+    Object.defineProperty(this, "_dataDir", { value: conf._dataDir });
+    Object.defineProperty(this, "_client", { value: conf._client });
     return this;
   }
 
@@ -32,7 +33,7 @@ class NumberConfig {
     if (this.min && parseInt(value) < this.min) return `Error while setting the value. ${value} is less than ${this.min}`;
     if (this.max && parseInt(value) > this.max) return `Error while setting the value. ${value} is more than ${this.max}`;
     this.data = value;
-    fs.outputJSONAsync(path.resolve(`${dataDir}${path.sep}${this._id}.json`), guildConfs.get(this._id));
+    fs.outputJSONAsync(path.resolve(`${this._dataDir}${path.sep}${this._id}.json`), this._client.guildConfs.get(this._id));
     return this;
   }
 
@@ -43,7 +44,7 @@ class NumberConfig {
    */
   setMin(value) {
     this.min = value;
-    fs.outputJSONAsync(path.resolve(`${dataDir}${path.sep}${this._id}.json`), guildConfs.get(this._id));
+    fs.outputJSONAsync(path.resolve(`${this._dataDir}${path.sep}${this._id}.json`), this._client.guildConfs.get(this._id));
     return this;
   }
 
@@ -54,7 +55,7 @@ class NumberConfig {
    */
   setMax(value) {
     this.max = value;
-    fs.outputJSONAsync(path.resolve(`${dataDir}${path.sep}${this._id}.json`), guildConfs.get(this._id));
+    fs.outputJSONAsync(path.resolve(`${this._dataDir}${path.sep}${this._id}.json`), this._client.guildConfs.get(this._id));
     return this;
   }
 }